Update confetti.browser.min.js
Here's the Xmas Edition of the confetti tool!

I did a bit un-minify to allow better editing. In this first sprint, the shapes set for Xmas is hard coded. On may refactor to externalize this later.

The code change is taken from the upstream pull request https://github.com/catdad/canvas-confetti/pull/111 , which is not processed for now. Because it's not restricted to emojis but may be used with any text, I changed the token used for declaration.