Update custom_ofmeet.js
Use Credentials API instead of Local Store to persist the XMPP credentials -- if available ike for Chrome, but still not for FireFox.

This still use `xmpp_{username,password}_override` in Local Store to pass the credentials to the Jitsi connect facility, but just for a short time during connection setup. This will avoid to patch the upstream code.