Update custom_ofmeet.js
If Conference Captions is enabled, all chat messages are duplicated to the lower main screen to get attention of the participants. If the chat message is a longer one, this may disturb the free view to the main window.

Introduce a configurable timeout that auto-clears the display of the message.